configure.ac modified to not link to SDL_ttf when using SDL_Pango
git-svn-id: svn://svn.debian.org/svn/tux4kids/tuxtype/trunk@931 47d30e19-490b-0410-9d18-e851f4e465b3

+++ src/playgame.c
@@ -1527,8 +1527,8 @@ static void SpawnFishies(int diflevel, int* fishies, int* frame)
   {
     /* NOTE need %S rather than %s because of wide characters */
     fprintf(stderr, "Spawn fishy with word '%S'\n", fish_object[*fishies].word);
-    fprintf(stderr, "Byte length is: %d\n", wcslen(fish_object[*fishies].word));
-    fprintf(stderr, "UTF-8 char length is: %d\n", fish_object[*fishies].len);
+    fprintf(stderr, "Byte length is: %d\n", (int)wcslen(fish_object[*fishies].word));
+    fprintf(stderr, "UTF-8 char length is: %d\n", (int)fish_object[*fishies].len);
   }
 
   *fishies = *fishies + 1;
